feat: implement Pengaduan Khusus feature with CRUD operations, export functionality, and PDF generation
- Added columns definition for displaying Pengaduan Khusus data in a table format.
- Created a form for submitting new Pengaduan Khusus entries.
- Implemented export functionality for Pengaduan Khusus data to PDF.
- Developed a detailed view for individual Pengaduan Khusus entries.
- Added status update capabilities for Pengaduan Khusus.
- Designed a PDF template for exporting Pengaduan Khusus reports.

+++ routes/web.php
@@ -6,48 +6,64 @@ use App\Models\Pengaduan;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\HomeController;
+use App\Http\Controllers\UserController;
+use App\Services\Dashboard\DashboardService;
 use App\Http\Controllers\PengaduanController;
 use App\Http\Controllers\ExportPengaduanController;
+use App\Http\Controllers\PengaduanKhususController;
 use App\Http\Controllers\TanggapanPengaduanController;
-use App\Http\Controllers\UserController;
-use App\Http\Resources\AuthUserResource;
 
 Route::get('/', [HomeController::class, 'index'])->name('home');
 
 Route::middleware('auth')->group(function () {
-    Route::resource('pengaduan', PengaduanController::class);
-    Route::put('pengaduan/update-validasi-rektor/{pengaduan}', [PengaduanController::class, 'updateValidasi'])->name('pengaduan.update.validasi');
+    Route::resource('umum/pengaduan', PengaduanController::class)
+        ->middlewareFor(['create', 'store'], 'role:mahasiswa')
+        ->middlewareFor(['destroy', 'update'], 'role:admin');
+    Route::put('pengaduan/update-validasi-rektor/{pengaduan}', [PengaduanController::class, 'updateValidasi'])->name('pengaduan.update.validasi')->middleware('role:rektor');
     Route::resource('tanggapan', TanggapanPengaduanController::class);
-    Route::get('/export-pdf', [ExportPengaduanController::class, 'export']);
-    Route::get('/users', [UserController::class, 'index'])->name('daftar.user');
-    Route::delete('/users/{user}', [UserController::class, 'destroy'])->name('delete.user');
+    Route::get('/export-pdf', [ExportPengaduanController::class, 'export'])->middleware('role:admin|warek_1|warek_2|warek_3|rektor');
+    Route::get('/export-pdf-pengaduan-khusus', [ExportPengaduanController::class, 'exportPengaduanKhusus'])->middleware('role:admin|warek_1|warek_2|warek_3|rektor');
+
+    Route::get('/users', [UserController::class, 'index'])->name('daftar.user')->middleware('role:admin');
+    Route::delete('/users/{user}', [UserController::class, 'destroy'])->name('delete.user')->middleware('role:admin');
+
+    Route::resource('khusus/pengaduan-khusus', PengaduanKhususController::class)
+        ->parameters(['pengaduan-khusus' => 'pengaduanKhusus'])
+        ->middlewareFor(['create', 'store'], 'role:mahasiswa')
+        ->middlewareFor(['destroy', 'update'], 'role:admin');
+    Route::put('pengaduan/update-validasi-rektor/{pengaduanKhusus}', [PengaduanKhususController::class, 'updateValidasi'])->name('pengaduan.update.validasi')->middleware('role:rektor');
 });
 
 Route::middleware(['auth', 'verified'])->group(function () {
-    Route::get('dashboard', function () {
+    Route::get('dashboard', function (DashboardService $dashboardService) {
         /** @var \App\Models\User */
         $user = Auth::user();
+        $role = $user->getRoleNames()->first();
 
-        if ($user->hasRole(['admin', 'rektor'])) {
-            $countAllPengaduan = Pengaduan::count();
-            $countPengaduanDiajukan = Pengaduan::denganStatus('Diajukan')->count();
-            $countPengaduanDiproses = Pengaduan::denganStatus('Diproses')->count();
-            $countPengaduanSelesai = Pengaduan::denganStatus('Selesai')->count();
-            $chartData = Pengaduan::chartData()->get();
+        $viewMahasiswa = 'menu/dashboard/mahasiswa/index';
+        $viewAdmin = 'menu/dashboard/admin/dashboard';
 
-            $pengaduanTervalidasi = Pengaduan::latest()->take(5)->get();
-
-            return Inertia::render('menu/dashboard/admin/dashboard', compact('countAllPengaduan', 'countPengaduanDiajukan', 'countPengaduanDiproses', 'countPengaduanSelesai', 'chartData', 'pengaduanTervalidasi'));
-        } elseif ($user->hasRole('mahasiswa')) {
-            $countAllPengaduan = Pengaduan::where('user_id', $user->id)->count();
-            $countPengaduanDiajukan = Pengaduan::where('user_id', $user->id)->denganStatus('Diajukan')->count();
-            $countPengaduanDiproses = Pengaduan::where('user_id', $user->id)->denganStatus('Diproses')->count();
-            $countPengaduanSelesai = Pengaduan::where('user_id', $user->id)->denganStatus('Selesai')->count();
-
-            return Inertia::render('menu/dashboard/mahasiswa/index', compact('countAllPengaduan', 'countPengaduanDiajukan', 'countPengaduanDiproses', 'countPengaduanSelesai'));
+        if ($role === 'mahasiswa') {
+            return Inertia::render($viewAdmin, $dashboardService->getDashboardData(userId: $user->id));
         }
 
-        return Inertia::render('dashboard');
+        $bidangMapping = [
+            'warek_1' => 'Akademik',
+            'warek_2' => 'Keuangan dan Umum',
+            'warek_3' => 'Kemahasiswaan',
+        ];
+
+        $bidang = $bidangMapping[$role] ?? null;
+
+        if (in_array($role, ['admin', 'rektor'])) {
+            return Inertia::render($viewAdmin, $dashboardService->getDashboardData());
+        }
+
+        if ($bidang) {
+            return Inertia::render($viewAdmin, $dashboardService->getDashboardData(bidang: $bidang));
+        }
+
+        abort(403);
     })->name('dashboard');
 });
